requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
aiohttp==3.11.11
aiosqlite==0.20.0
requests==2.31.0
